backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-multipart
pdfplumber
python-docx
google-generativeai
playwright
celery
redis
email-validator
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
axios
